Added splash screen and Modified index
