+++ package.json
@@ -11,18 +11,21 @@
   "dependencies": {
     "@gorhom/bottom-sheet": "^4.6.0",
     "@react-native-community/slider": "^4.5.0",
-    "expo": "^50.0.1",
+    "expo": "^50.0.7",
     "expo-constants": "~15.4.2",
     "expo-font": "^11.10.3",
     "expo-linking": "~6.2.1",
     "expo-router": "~3.4.1",
     "expo-splash-screen": "~0.26.1",
     "expo-status-bar": "~1.11.1",
+    "lottie-ios": "^3.2.3",
+    "lottie-react-native": "6.5.1",
     "nativewind": "^4.0.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native": "0.73.2",
     "react-native-gesture-handler": "~2.14.0",
+    "react-native-lottie-splash-screen": "^1.1.2",
     "react-native-maps": "^1.10.3",
     "react-native-reanimated": "~3.6.0",
     "react-native-safe-area-context": "4.8.2",
